Adjust the Folder Structure and set up the navigation for frontend
