+++ app/src/main/java/com/example/calorietrackerapp/MainActivity.kt
@@ -49,6 +49,11 @@ import androidx.navigation.NavController
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
+import com.example.calorietrackerapp.ui.screens.AuthLandingScreen
+import com.example.calorietrackerapp.ui.screens.HomeScreen
+import com.example.calorietrackerapp.ui.screens.MealRecord
+import com.example.calorietrackerapp.ui.screens.RegisterScreen
+import com.example.calorietrackerapp.ui.screens.SearchMeal
 import com.example.calorietrackerapp.ui.theme.CalorieTrackerAppTheme
 
 
@@ -60,246 +65,14 @@ class MainActivity : ComponentActivity() {
             CalorieTrackerAppTheme {
                 val navController = rememberNavController()
 
-                NavHost(navController = navController, startDestination = "main") {
-                    composable("main") { MainScreen(navController) }
+                NavHost(navController = navController, startDestination = "home") {
+                    composable("authLanding") { AuthLandingScreen(navController) }
                     composable("register") { RegisterScreen(navController) }
                     composable("home") { HomeScreen(navController)}
+                    composable("searchMeal") { SearchMeal(navController) }
+                    composable("mealRecord") { MealRecord(navController) }
+
                 }
-            }
-        }
-    }
-
-    @Composable
-    fun MainScreen(navController: NavController) {
-        Box(
-            modifier = Modifier
-                .fillMaxSize()
-                .background(Color.White)
-        ) {
-            Column(
-                modifier = Modifier.fillMaxSize()
-            ) {
-                // makes the green curve of the welcome screen
-                Box(
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .height(300.dp)
-                        .clip(RoundedCornerShape(bottomStart = 150.dp, bottomEnd = 150.dp))
-                        .background(Color(0xFF55745D))
-                )
-
-                Spacer(modifier = Modifier.weight(1f)) // pushes buttons to bottom
-            }
-
-
-            Row(
-                modifier = Modifier
-                    .align(Alignment.BottomCenter)
-                    .padding(bottom = 64.dp, start = 16.dp, end = 16.dp),
-                horizontalArrangement = Arrangement.spacedBy(12.dp)
-            ) {
-                // sign in button
-                Button(
-                    onClick = { /* logic TBA */ },
-                    modifier = Modifier
-                        .weight(1f)
-                        .height(56.dp),
-                    shape = RoundedCornerShape(16.dp),
-                    colors = ButtonDefaults.buttonColors(containerColor = Color(0xFF55745D))
-                ) {
-                    Text(
-                        "SIGN IN",
-                        color = Color.White,
-                        fontSize = 19.sp,
-                        fontWeight = FontWeight.Bold
-                    )
-                }
-
-                // register button
-                Button(
-                    onClick = { navController.navigate("register")},
-                    modifier = Modifier
-                        .weight(1f)
-                        .height(56.dp),
-                    shape = RoundedCornerShape(16.dp),
-                    colors = ButtonDefaults.buttonColors(containerColor = Color(0xFFE8EBE8))
-                ) {
-                    Text(
-                        "REGISTER",
-                        color = Color(0xFF55745D),
-                        fontSize = 19.sp,
-                        fontWeight = FontWeight.Bold
-                    )
-                }
-            }
-        }
-    }
-
-    @Composable
-        fun RegisterScreen(navController: NavController) {
-            var username by remember { mutableStateOf("") }
-            var password by remember { mutableStateOf("") }
-
-            Box(
-                modifier = Modifier
-                    .fillMaxSize()
-                    .background(Color(0xFFE8EBE8))
-                    .padding(32.dp),
-                contentAlignment = Alignment.Center
-            ) {
-                Column(
-                    verticalArrangement = Arrangement.spacedBy(16.dp),
-                    horizontalAlignment = Alignment.CenterHorizontally
-                ) {
-                    Text("Register", fontSize = 28.sp, fontWeight = FontWeight.Bold)
-
-                    // input username logic
-                    OutlinedTextField(
-                        value = username,
-                        onValueChange = { username = it },
-                        label = { Text("Username") }
-                    )
-
-                    // input password logic
-                    OutlinedTextField(
-                        value = password,
-                        onValueChange = { password = it },
-                        label = { Text("Password") },
-                        visualTransformation = PasswordVisualTransformation()
-                    )
-
-                    // Register button
-                    Button(
-                        onClick = {
-                            navController.navigate("home")
-                        },
-                        modifier = Modifier.fillMaxWidth(),
-                        colors = ButtonDefaults.buttonColors(containerColor = Color(0xFF55745D))
-                    ) {
-                        Text("Register")
-                    }
-
-                    // Back button logic
-                    TextButton(onClick = { navController.popBackStack() }, // popBackStack goes to the previous screen
-                        colors = ButtonDefaults.textButtonColors(contentColor = Color(0xFF55745D))
-                    ) {
-                        Text("Back")
-                    }
-                }
-            }
-        }
-
-
-    @Composable
-    fun HomeScreen(navController: NavController) {
-
-        val viewModel: MealSearchViewModel = viewModel()
-        val results by viewModel.searchResults.collectAsState()
-
-        var searchQuery by remember { mutableStateOf("")}
-
-        Box(
-            modifier = Modifier
-                .fillMaxSize()
-                .background(Color.White)
-        ) {
-            Column(
-                modifier = Modifier
-                    .fillMaxSize()
-                    .padding(16.dp)
-            ) {
-
-                Spacer(modifier = Modifier.height(20.dp))
-
-                // Today header
-                Text(
-                    text = "Today",
-                    fontSize = 28.sp,
-                    fontWeight = FontWeight.Bold
-                )
-
-                Spacer(modifier = Modifier.height(20.dp))
-
-                // summary of calories
-                Box(
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .height(150.dp)
-                        .background(Color(0xFF55745D))
-                        .padding(16.dp)
-                ) {
-                    Text(
-                        text = "2078 Calories Left",
-                        color = Color.White,
-                        fontSize = 20.sp,
-                        fontWeight = FontWeight.Bold,
-                        modifier = Modifier.align(Alignment.Center)
-                    )
-                }
-
-                Spacer(modifier = Modifier.height(24.dp))
-
-                // search meal box
-                Row(
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .height(70.dp)
-                        .clip(RoundedCornerShape(12.dp))
-                        .background(Color(0xFFF5F5F5))
-                        .padding(horizontal = 8.dp, vertical = 4.dp),
-                    verticalAlignment = Alignment.CenterVertically
-                ) {
-                    OutlinedTextField(
-                        value = searchQuery,
-                        onValueChange = {
-                            searchQuery = it
-                            if (it.isNotBlank()) viewModel.searchMeal(it)
-                        },
-                        placeholder = { Text("Find your meal's calories here...") },
-                        modifier = Modifier.weight(1f),
-                        maxLines = 1,
-                        singleLine = true
-                    )
-                    IconButton(onClick = { /* TODO: Search Action */ }) {
-                        Icon(Icons.Default.Search, contentDescription = "Search")
-                    }
-                }
-
-                Spacer(modifier = Modifier.height(24.dp))
-
-                // Today's Meals Title
-                Text(
-                    text = "Today's Meals",
-                    fontSize = 20.sp,
-                    fontWeight = FontWeight.Bold,
-                    modifier = Modifier.padding(bottom = 8.dp)
-                )
-
-                // Meal List
-                LazyColumn(
-                    verticalArrangement = Arrangement.spacedBy(8.dp)
-                ) {
-                    items(results) { meal ->
-                        Text(
-                            text = "${meal.name}: ${meal.calories} Calories",
-                            modifier = Modifier.padding(8.dp)
-                        )
-                    }
-                }
-            }
-
-            // Bottom Navigation Bar
-            Row(
-                modifier = Modifier
-                    .align(Alignment.BottomCenter)
-                    .fillMaxWidth()
-                    .background(Color(0xFFF5F5F5))
-                    .padding(8.dp),
-                horizontalArrangement = Arrangement.SpaceEvenly
-            ) {
-                TextButton(onClick = { /* TBA */ }) { Text("Today's Calories") }
-                TextButton(onClick = { /* TBA */ }) { Text("Search Meal") }
-                TextButton(onClick = { /* TBA */ }) { Text("Meal Planning") }
             }
         }
     }
