+++ app/src/main/java/com/example/calorietrackerapp/ui/screens/components/NavBar.kt
@@ -1,0 +1,38 @@
+package com.example.calorietrackerapp.ui.screens.components
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.Text
+import androidx.compose.material3.TextButton
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.unit.dp
+import androidx.navigation.NavController
+
+
+@Composable
+fun NavBar(navController: NavController){
+    Box(modifier = Modifier.fillMaxSize()){
+        // Bottom Navigation Bar
+        Row(
+            modifier = Modifier
+                .align(Alignment.BottomCenter)
+                .fillMaxWidth()
+                .background(Color(0xFFF5F5F5))
+                .padding(8.dp),
+            horizontalArrangement = Arrangement.SpaceEvenly
+        ) {
+            TextButton(onClick = { navController.navigate("home") }) { Text("Today's Calories") }
+            TextButton(onClick = { navController.navigate("searchMeal") }) { Text("Search Meal") }
+            TextButton(onClick = { navController.navigate("mealRecord") }) { Text("Meal Record") }
+        }
+    }
+
+}
